refactor(god_object): extrai stage handlers do PipelineMonitor
PipelineMonitor (~962 linhas) misturava lifecycle, scheduling e sete
stage handlers. Os handlers foram movidos para
deile/orchestration/pipeline/stages.py como funcoes livres que recebem
o monitor; PipelineMonitor mantem lifecycle + scheduling e delega.
